REF: Change layers so that they are as per wiki
