+++ android/feature/rocket/src/test/java/cz/quanti/rocketapp/android/feature/rocket/presentation/RocketDetailViewModelTest.kt
@@ -1,11 +1,11 @@
 package cz.quanti.rocketapp.android.feature.rocket.presentation
 
-import cz.quanti.rocketapp.android.feature.rocket.util.rocketsDetailData
+import cz.quanti.rocketapp.android.feature.rocket.infrastructure.rocketsDetailData
 import cz.quanti.rocketapp.android.lib.uisystem.presentation.UiScreenState
 import cz.quanti.rocketapp.android.lib.uisystem.presentation.UiText
 import cz.quanti.rocketapp.android.rocket.R
 import cz.quanti.rocketapp.multiplatform.feature.rocket.domain.GetRocketUseCase
-import cz.quanti.rocketapp.multiplatform.lib.common.asResult
+import cz.quanti.rocketapp.multiplatform.lib.common.infrastructure.asResult
 import io.kotest.matchers.shouldBe
 import org.junit.After
 import org.junit.Before
