+++ android/lib/architecture-test/src/test/java/cz/quanti/rocketapp/android/lib/architecturetest/infrastructure/KonsistExtensions.kt
@@ -1,12 +1,12 @@
-package cz.quanti.rocketapp.android.lib.architecturetest.util
+package cz.quanti.rocketapp.android.lib.architecturetest.infrastructure
 
 import com.lemonappdev.konsist.api.declaration.KoFileDeclaration
 import com.lemonappdev.konsist.api.provider.KoHasPackageProvider
 import com.lemonappdev.konsist.api.provider.KoResideInPackageProvider
-import cz.quanti.rocketapp.android.lib.architecturetest.util.model.ArchitectureLayer
-import cz.quanti.rocketapp.android.lib.architecturetest.util.model.ModuleType
-import cz.quanti.rocketapp.android.lib.architecturetest.util.model.PackagePatterns
-import cz.quanti.rocketapp.android.lib.architecturetest.util.model.Platform
+import cz.quanti.rocketapp.android.lib.architecturetest.infrastructure.model.ArchitectureLayer
+import cz.quanti.rocketapp.android.lib.architecturetest.infrastructure.model.ModuleType
+import cz.quanti.rocketapp.android.lib.architecturetest.infrastructure.model.PackagePatterns
+import cz.quanti.rocketapp.android.lib.architecturetest.infrastructure.model.Platform
 
 fun KoHasPackageProvider.resideInLayer(layer: ArchitectureLayer): Boolean {
     return hasPackage("${PackagePatterns.ROOT_PACKAGE}..${layer.layerName}..")
