+++ android/lib/architecture-test/src/test/java/cz/quanti/rocketapp/android/lib/architecturetest/infrastructure/model/PackagePatterns.kt
@@ -1,4 +1,4 @@
-package cz.quanti.rocketapp.android.lib.architecturetest.util.model
+package cz.quanti.rocketapp.android.lib.architecturetest.infrastructure.model
 
 import cz.quanti.rocketapp.android.architecturetest.BuildConfig
