+++ multiplatform/feature/rocket/src/commonMain/kotlin/cz/quanti/rocketapp/multiplatform/feature/rocket/domain/GetRocketUseCase.kt
@@ -2,8 +2,8 @@ package cz.quanti.rocketapp.multiplatform.feature.rocket.domain
 
 import com.rickclephas.kmp.nativecoroutines.NativeCoroutines
 import cz.quanti.rocketapp.multiplatform.feature.rocket.model.RocketDetail
-import cz.quanti.rocketapp.multiplatform.lib.common.Result
 import cz.quanti.rocketapp.multiplatform.lib.common.domain.FlowUseCase
+import cz.quanti.rocketapp.multiplatform.lib.common.model.Result
 import kotlinx.coroutines.flow.Flow
 
 fun interface GetRocketUseCase : FlowUseCase<String, Result<RocketDetail>> {
