+++ multiplatform/feature/rocket/src/commonMain/kotlin/cz/quanti/rocketapp/multiplatform/feature/rocket/domain/GetRocketsUseCaseImpl.kt
@@ -2,8 +2,8 @@ package cz.quanti.rocketapp.multiplatform.feature.rocket.domain
 
 import cz.quanti.rocketapp.multiplatform.feature.rocket.data.RocketData
 import cz.quanti.rocketapp.multiplatform.feature.rocket.model.Rocket
-import cz.quanti.rocketapp.multiplatform.lib.common.Result
-import cz.quanti.rocketapp.multiplatform.lib.common.asResult
+import cz.quanti.rocketapp.multiplatform.lib.common.infrastructure.asResult
+import cz.quanti.rocketapp.multiplatform.lib.common.model.Result
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 
@@ -11,9 +11,11 @@ internal class GetRocketsUseCaseImpl(
     private val repo: RocketsRepository
 ) : GetRocketsUseCase {
     override fun invoke(input: Unit): Flow<Result<List<Rocket>>> {
-        return repo.getRockets().map { list ->
-            list.map { it.asRocket() }
-        }.asResult()
+        return repo.getRockets()
+            .map { list ->
+                list.map { it.asRocket() }
+            }
+            .asResult()
     }
 }
