+++ multiplatform/lib/common/src/commonMain/kotlin/cz/quanti/rocketapp/multiplatform/lib/common/infrastructure/ResultExtension.kt
@@ -1,5 +1,7 @@
-package cz.quanti.rocketapp.multiplatform.lib.common
+package cz.quanti.rocketapp.multiplatform.lib.common.infrastructure
 
+import cz.quanti.rocketapp.multiplatform.lib.common.model.Result
+import cz.quanti.rocketapp.multiplatform.lib.common.model.ResultException
 import io.ktor.client.engine.ClientEngineClosedException
 import io.ktor.client.network.sockets.ConnectTimeoutException
 import io.ktor.client.network.sockets.SocketTimeoutException
@@ -13,15 +15,7 @@ import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.onStart
 
-sealed interface Result<out T> {
-    data class Success<T>(val data: T) : Result<T>
-
-    data class Error(val exception: ResultException? = null) : Result<Nothing>
-
-    data object Loading : Result<Nothing>
-}
-
-// TODO: flow on ios?
+@Suppress("LongMethod")
 fun <T> Flow<T>.asResult(): Flow<Result<T>> {
     return this
         .map<T, Result<T>> {
@@ -34,51 +28,57 @@ fun <T> Flow<T>.asResult(): Flow<Result<T>> {
                     emit(
                         Result.Error(
                             ResultException.NetworkException(
-                                it.message ?: "Network timeout", it
-                            )
-                        )
+                                it.message ?: "Network timeout", it,
+                            ),
+                        ),
                     )
+
                 is IOException ->
                     emit(
                         Result.Error(
                             ResultException.NetworkException(
-                                it.message ?: "Network error", it
-                            )
-                        )
+                                it.message ?: "Network error", it,
+                            ),
+                        ),
                     )
+
                 is ResponseException ->
                     emit(Result.Error(ResultException.HttpException(it.response.status, it)))
+
                 is SendCountExceedException ->
                     emit(
                         Result.Error(
                             ResultException.NetworkException(
-                                it.message ?: "Infinite or too long redirect", it
+                                it.message ?: "Infinite or too long redirect", it,
                             ),
                         ),
                     )
+
                 is ClientEngineClosedException ->
                     emit(
                         Result.Error(
                             ResultException.Exception(
-                                it.message ?: "Client engine closed", it
-                            )
-                        )
+                                it.message ?: "Client engine closed", it,
+                            ),
+                        ),
                     )
+
                 is ContentConverterException ->
                     emit(
                         Result.Error(
                             ResultException.ContentException(
-                                it.message ?: "Content conversion error", it
-                            )
-                        )
+                                it.message ?: "Content conversion error", it,
+                            ),
+                        ),
                     )
+
                 else -> emit(
                     Result.Error(
                         ResultException.Exception(
                             it.message ?: "Unknown error",
-                            it
-                        )
-                    )
+                            it,
+                        ),
+                    ),
                 )
             }
         }
