+++ multiplatform/lib/common/src/commonMain/kotlin/cz/quanti/rocketapp/multiplatform/lib/common/model/Result.kt
@@ -1,0 +1,9 @@
+package cz.quanti.rocketapp.multiplatform.lib.common.model
+
+sealed interface Result<out T> {
+    data class Success<T>(val data: T) : Result<T>
+
+    data class Error(val exception: ResultException? = null) : Result<Nothing>
+
+    data object Loading : Result<Nothing>
+}
